edit models/students.py for receive in grades list[float]

+++ models/students.py
@@ -4,13 +4,13 @@ class Student(BaseModel):
     id: str | None = None
     name: str
     age: int
-    grades: list[int]
+    grades: list[float]
     passed: bool
     
     
 class Student_grades(BaseModel):
     id: str
-    grades: list[int]
+    grades: list[float]
     
 class Student_passed(BaseModel):
     id: str
